Add files via upload
Primarily this represents a refactoring of the previous version; in particular, I added a Data_Processor to manage an object hierarchy built on classes that inherit from Data_Object.  Data_Object is a class that abstracts the hierarchical relationship, primarily maintaining is_dirty in relationship to other objects in the hierarchy. It provides a uniform interface for loading data Just In Time, using load_data() and ensure().